Disable forced casting to isize for enum

+++ binding-generator/src/writer/rust_native/tpl/enum/const.tpl.rs
@@ -1,3 +1,3 @@
 {{doc_comment}}
-{{name}} = {{value}} as isize,
+{{name}} = {{value}},
